+++ src/test/java/io/keyko/nevermined/api/AssetsApiIT.java
@@ -75,11 +75,12 @@ public class AssetsApiIT {
 
         String metadataUrl = config.getString("metadata-internal.url") + "/api/v1/metadata/assets/ddo/{did}";
         String provenanceUrl = config.getString("metadata-internal.url") + "/api/v1/metadata/assets/provenance/{did}";
-        String consumeUrl = config.getString("gateway.url") + "/api/v1/gateway/services/consume";
+        String gatewayUrl = config.getString("gateway.url");
+        String consumeUrl = gatewayUrl + "/api/v1/gateway/services/access";
         String secretStoreEndpoint = config.getString("secretstore.url");
         String providerAddress = config.getString("provider.address");
 
-        providerConfig = new ProviderConfig(consumeUrl, metadataUrl, provenanceUrl, secretStoreEndpoint, providerAddress);
+        providerConfig = new ProviderConfig(consumeUrl, metadataUrl, gatewayUrl, provenanceUrl, secretStoreEndpoint, providerAddress);
 
         neverminedAPI = NeverminedAPI.getInstance(config);